VideoRentalSystem/app/src/main/java/com/chrisneric/videorentalsystem/entity/Account.java
package com.chrisneric.videorentalsystem.entity;

import android.arch.persistence.room.ColumnInfo;
import android.arch.persistence.room.Entity;
import android.arch.persistence.room.PrimaryKey;

import static com.chrisneric.videorentalsystem.database.AppDatabase.getAccountById;

@Entity(tableName = "account")
public class Account {
    @PrimaryKey(autoGenerate = true)
    private int uid;
    @ColumnInfo(name = "name")
    private String name;
    @ColumnInfo(name = "address")
    private String address;
    @ColumnInfo(name = "email")
    private String email;
    @ColumnInfo(name = "phone")
    private String phone;

    public Account() {

    }

    public Account(String n, String a, String e, String p) {
        name = n;
        address = a;
        email = e;
        phone = p;
    }

    public static Account lookupById(int i) {
        return getAccountById(i);
    }

    public String getAddress() {
        return address;
    }

    public void setAddress(String address) {
        this.address = address;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPhone() {
        return phone;
    }

    public void setPhone(String phone) {
        this.phone = phone;
    }

    public int getUid() {
        return uid;
    }

    public void setUid(int uid) {
        this.uid = uid;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}
